상품등록 - 상품리스트

+++ src/main/java/com/example/rewrite/entity/Product.java
@@ -18,7 +18,7 @@ public class Product {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name="PROD_ID",nullable = false ,length = 255)
-    private int prodId;
+    private Long prodId;
 
     @Column(name = "PRICE")
     private String price;
